Extract LongPressDetector into a separate file. Make the UI tidier during power off.
